requirements.txt
python-pptx
Pillow
face_recognition
